ResultPage DateService Update

+++ src/main/java/com/douzone/dataservice/domain/ResultDTO.java
@@ -5,16 +5,16 @@ import java.sql.Date;
 
 @Data
 public class ResultDTO {
-    private int resultNo;
-    private int patientNo;
+    private Integer resultNo;
+    private Integer patientNo;
     private String patientName;
     private String patientPhoneNumber;
     private String registerDt;
     private String sampleName;
     private String prescribeDt;
     private String inspectionName;
-    private int figures;
-    private int baseline;
+    private Long figures;
+    private Integer baseline;
     private String unit;
     private String note;
 }
